+++ src/util.cpp
@@ -106,8 +106,9 @@ void Log::alert(const std::string &msg, AlertDetails details) {
 void Log::slack(const std::string &msg, AlertDetails details) {
   std::string cmd;
   if (!details.text.empty()) {
-    std::replace(details.text.begin(), details.text.end(), '"', ' ');
-    std::replace(details.title.begin(), details.title.end(), '"', ' ');
+    replaceAll(details.text, "\\", "\\\\");  // first escape backslashes
+    replaceAll(details.text, "\"", "\\\"");  // then the rest
+    replaceAll(details.title, "\"", "\\\"");
     size_t index = 0;
     while (true) {
       index = details.text.find("$", index);